requirement.txt
langchain-google-genai
langchain-core
langgraph
python-dotenv
pandas
tabulate
yfinance
selenium
